+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing yet
+### Added
+
+- InitSegment.TweakSingleTrakLive changes an init segment to fit live streaming
 
 ## [0.42.0] - 2024-01-26